+++ tests/Integration/Integrations/Curl/CurlIntegrationTest.php
@@ -2,9 +2,13 @@
 
 namespace DDTrace\Tests\Integration\Integrations\Curl;
 
+use DDTrace\Configuration;
+use DDTrace\Formats;
 use DDTrace\Integrations\Curl\CurlIntegration;
 use DDTrace\Tests\Integration\Common\IntegrationTestCase;
 use DDTrace\Tests\Integration\Common\SpanAssertion;
+use DDTrace\Util\ArrayKVStore;
+use OpenTracing\GlobalTracer;
 
 
 final class CurlIntegrationTest extends IntegrationTestCase
@@ -139,5 +143,64 @@ final class CurlIntegrationTest extends IntegrationTestCase
                 ])
                 ->setError(),
         ]);
+    }
+
+    public function testKVStoreIsCleanedOnCurlClose()
+    {
+        $ch = curl_init(self::URL . '/status/200');
+        curl_setopt($ch, CURLOPT_HTTPHEADER, []);
+        $this->assertNotSame('default', ArrayKVStore::getForResource($ch, Formats\CURL_HTTP_HEADERS, 'default'));
+        curl_close($ch);
+        $this->assertSame('default', ArrayKVStore::getForResource($ch, Formats\CURL_HTTP_HEADERS, 'default'));
+    }
+
+    public function testDistributedTracingIsPropagated()
+    {
+        $found = [];
+
+        $traces = $this->isolateTracer(function () use (&$found) {
+
+            $tracer = GlobalTracer::get();
+            $span = $tracer->startActiveSpan('some_operation')->getSpan();
+
+            $ch = curl_init(self::URL . '/headers');
+            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+            curl_setopt($ch, CURLOPT_HTTPHEADER, [
+                'honored: preserved_value',
+            ]);
+            $found = json_decode(curl_exec($ch), 1);
+
+            $span->finish();
+        });
+
+        // trace is: some_operation
+        $this->assertSame($traces[0][0]->getContext()->getSpanId(), $found['headers']['X-Datadog-Trace-Id']);
+        // parent is: curl_exec
+        $this->assertSame($traces[0][1]->getContext()->getSpanId(), $found['headers']['X-Datadog-Parent-Id']);
+        // existing headers are honored
+        $this->assertSame('preserved_value', $found['headers']['Honored']);
+    }
+
+    public function testDistributedTracingIsNotPropagatedIfDisabled()
+    {
+        $found = [];
+        Configuration::replace(\Mockery::mock('\DDTrace\Configuration', [
+            'isDistributedTracingEnabled' => false
+        ]));
+
+        $this->isolateTracer(function () use (&$found) {
+
+            $tracer = GlobalTracer::get();
+            $span = $tracer->startActiveSpan('some_operation')->getSpan();
+
+            $ch = curl_init(self::URL . '/headers');
+            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+            $found = json_decode(curl_exec($ch), 1);
+
+            $span->finish();
+        });
+
+        $this->assertArrayNotHasKey('X-Datadog-Trace-Id', $found['headers']);
+        $this->assertArrayNotHasKey('X-Datadog-Parent-Id', $found['headers']);
     }
 }